Support 202111 (#1101)
* Block registering private ASN ranges

* Add a continental region field for facilities #1007

* Incorrect order of search results #232

* Allow users to upload a small logo to their record #346

* Sponsor badge CSS and text translation issues #453

* IXP and Facility summary #18

* Add sales email and phone contact to ix object #949

* Increase timeout timer for IX-F JSON importer to 30s #962

* IX-F manually triggered import bugs #1054

* remove survey notifications

* IX-F importer: manual add followed by IX-F prompted add can result in "The server rejected your data" #807

* performance fix for net view

* api documentation generate broken #956

* fix poc set missing from net view

* Auto focus cursor on search field on main website #695

* PeeringDB website has a poor choice of line-breaks for IPv6 addresses. #748

* Delete childless org objects #838

* linting

* poetry lock

* re-add mistakingly dropped translation to email template

* fix template variables

* regen docs

* regen api docs

* add u2f api

* fix merge cruft

* fido u2f webauthn first steps

* dont need this

* typo

* first ux steps for security key management, u2f views

* u2f auth

* linting

* Cascade delete when performed by superuser in django-admin #735

* security key passwordless login

* reset webauthn migrations

* security key 2fa

* linting

* ux cleanup

* cleanup

* Add IXP to AS record / dropdown limited #469

* ignore orgs that have pre-dated or finished sponsorships attached

* split out to django_security_keys

* default to 90 days

* make URL required for new objects #374

* move CONN_MAX_AGE to proper location and expose as env var (#1060)

* Error message for invalid password on email change

* Registering a new facility or exchange organization is broken #1062

* Creating a facility that matches the name of a soft-deleted facility will cause the entry to bypass the verification queue #901

* irr source validator doesn't allow for hyphens in source #921

* split out javascript

* split out js

* Clicking "Add" to add a user api-key without providing a name for the key raises Internal Error #1033

* remove security key admin views

* global stats don't show up at login screen #284

* Possible for "pending" exchange to have "deleted" ixlan (#1077)

* webauthn finalizations

* oidc support for oauth provider #1070

* Revert "Merge branch 'gh_1070' into 'support_202111'"

* remove unused dev mounts

* linting

* add django-security-key and relock poetry

* re-add imports removed through linting

* docs regen

* oauth toolkit pinned to 1.6.1
django-security-keys pinned to 1.0.1

* remove debug message

* linting

* linting

+++ peeringdb_server/admin.py
@@ -338,24 +338,38 @@ rollback.short_description = _("ROLLBACK")
 @transaction.atomic
 @reversion.create_revision()
 def soft_delete(modeladmin, request, queryset):
-    if request.user:
-        reversion.set_user(request.user)
 
-    if queryset.model.handleref.tag == "ixlan":
-        messages.error(
-            request,
-            _(
-                "Ixlans can no longer be directly deleted as they are now synced to the parent exchange"
-            ),
+    if request.POST.get("delete"):
+
+        if request.user:
+            reversion.set_user(request.user)
+
+        if queryset.model.handleref.tag == "ixlan":
+            messages.error(
+                request,
+                _(
+                    "Ixlans can no longer be directly deleted as they are now synced to the parent exchange"
+                ),
+            )
+            return
+
+        for row in queryset:
+            try:
+                row.delete()
+            except ProtectedAction as err:
+                messages.error(request, _("Protected object '{}': {}").format(row, err))
+                continue
+    else:
+
+        context = dict(
+            admin.site.each_context(request),
+            deletable_objects=queryset,
+            action_checkbox_name=helpers.ACTION_CHECKBOX_NAME,
+            title=_("Delete selected objects"),
         )
-        return
 
-    for row in queryset:
-        try:
-            row.delete()
-        except ProtectedAction as err:
-            messages.error(request, _("Protected object '{}': {}").format(row, err))
-            continue
+        messages.warning(request, _("Please confirm deletion of selected objects."))
+        return TemplateResponse(request, "admin/soft_delete.html", context)
 
 
 soft_delete.short_description = _("SOFT DELETE")
@@ -497,13 +511,11 @@ class SoftDeleteAdmin(
     """
 
     actions = [soft_delete]
+
     object_history_template = "handleref/grappelli/object_history.html"
     version_details_template = "handleref/grappelli/version_details.html"
     version_revert_template = "handleref/grappelli/version_revert.html"
     version_rollback_template = "handleref/grappelli/version_rollback.html"
-
-    def has_delete_permission(self, request, obj=None):
-        return False
 
     @transaction.atomic
     @reversion.create_revision()
@@ -514,6 +526,21 @@ class SoftDeleteAdmin(
 
     def grainy_namespace(self, obj):
         return obj.grainy_namespace
+
+    def get_actions(self, request):
+        actions = super().get_actions(request)
+        if "delete_selected" in actions:
+            del actions["delete_selected"]
+        return actions
+
+
+class ProtectedDeleteAdmin(admin.ModelAdmin):
+    """
+    Allow deletion of objects if the user is superuser
+    """
+
+    def has_delete_permission(self, request, obj=None):
+        return request.user.is_superuser
 
 
 class ModelAdminWithVQCtrl:
@@ -736,9 +763,9 @@ class InternetExchangeAdmin(ModelAdminWithVQCtrl, SoftDeleteAdmin):
         "proto_multicast",
     )
 
-    raw_id_fields = ("org",)
+    raw_id_fields = ("org", "ixf_import_request_user")
     autocomplete_lookup_fields = {
-        "fk": ["org"],
+        "fk": ["org", "ixf_import_request_user"],
     }
 
     def ixf_import_history(self, obj):
@@ -1441,7 +1468,7 @@ class VerificationQueueAdmin(ModelAdminWithUrlActions):
         return ""
 
 
-class UserOrgAffiliationRequestAdmin(ModelAdminWithUrlActions):
+class UserOrgAffiliationRequestAdmin(ModelAdminWithUrlActions, ProtectedDeleteAdmin):
     list_display = (
         "user",
         "asn",
@@ -1499,9 +1526,6 @@ class UserOrgAffiliationRequestAdmin(ModelAdminWithUrlActions):
             each.deny()
 
     deny.short_description = _("Deny")
-
-    def has_delete_permission(self, request, obj=None):
-        return False
 
     actions = [approve_and_notify, approve, deny]
